+++ ui/components/providers/forms/muted-findings-config-form.tsx
@@ -16,7 +16,11 @@ import { CustomButton } from "@/components/ui/custom";
 import { CustomLink } from "@/components/ui/custom/custom-link";
 import { FormButtons } from "@/components/ui/form";
 import { fontMono } from "@/config/fonts";
-import { convertToYaml, parseYamlValidation } from "@/lib/yaml";
+import {
+  convertToYaml,
+  defaultMutedFindingsConfig,
+  parseYamlValidation,
+} from "@/lib/yaml";
 import {
   MutedFindingsConfigActionState,
   ProcessorData,
@@ -178,7 +182,7 @@ export const MutedFindingsConfigForm = ({
             </li>
             <li>
               Learn more about configuring the Mutelist{" "}
-              <CustomLink href="https://docs.prowler.com/projects/prowler-open-source/en/latest/tutorials/mutelist/">
+              <CustomLink href="https://docs.prowler.com/projects/prowler-open-source/en/latest/tutorials/prowler-app-mute-findings">
                 here
               </CustomLink>
               .
@@ -201,7 +205,7 @@ export const MutedFindingsConfigForm = ({
             <Textarea
               id="configuration"
               name="configuration"
-              placeholder="Enter your YAML configuration..."
+              placeholder={defaultMutedFindingsConfig}
               variant="bordered"
               value={configText}
               onChange={(e) => handleConfigChange(e.target.value)}